% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posterior.R
\name{get_posterior}
\alias{get_posterior}
\title{Compute Posterior Distribution}
\usage{
get_posterior(successes, trials, alpha_prior = 1, beta_prior = 1)
}
\arguments{
\item{successes}{number of success from binomial process}

\item{trials}{number of trials}

\item{alpha_prior}{alpha parameter for Beta distribution prior. Defaults to 1.}

\item{beta_prior}{beta parameter for Beta distribution prior. Defaults to 1.}
}
\description{
Computes the posterior distribution given the supplied data and Beta distirbution
priors.
}
\examples{
get_posterior( successes = 5, trials = 10 )
get_posterior( successes = 5, trials = 10, alpha_prior = 2, beta_prior = 2)
}
